Getting the values set up
Just trying to run this made me realize that if I am calling define, or allocate global arrays for the number of customers, but I want to do dynamic sizes for these things, this isn't gonna work. I can't overwrite a defined int, and I can't have this uninitialized global arrays. I am not sure why this is provided in the assignment code yet completely doesn't work with the assignment???